+++ mortar/src/main/java/mortar/MortarScope.java
@@ -32,9 +32,9 @@ public interface MortarScope {
   ObjectGraph getObjectGraph();
 
   /**
-   * Register the given {@link Scoped} instance to have its {@link Scoped#onDestroy()} method
-   * called from {@link #destroy()}. Redundant calls are safe, they will not lead to double
-   * registration.
+   * Register the given {@link Scoped} instance to have its {@link Scoped#onExitScope()}
+   * method called from {@link MortarScope#destroy}. Redundant calls are safe, they will not lead to
+   * double registration.
    *
    * @throws IllegalStateException if this scope has been destroyed
    */
@@ -67,9 +67,9 @@ public interface MortarScope {
   Context createContext(Context parentContext);
 
   /**
-   * Sends {@link Scoped#onDestroy()} to all registrants and then clears the registration
-   * list. Recursively destroys all children. Parent scope drops its reference to this instance.
-   * Redundant calls to this method are safe.
+   * Sends {@link Scoped#onExitScope()} to all registrants and then clears the
+   * registration list. Recursively destroys all children. Parent scope drops its reference
+   * to this instance. Redundant calls to this method are safe.
    */
-  void destroy();
+  void destroyChild(MortarScope child);
 }