+++ mortar/src/main/java/mortar/RealScope.java
@@ -25,14 +25,17 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import static java.lang.Integer.toHexString;
+import static java.lang.String.format;
+
 class RealScope implements MortarScope {
 
   protected final boolean validate;
-  protected final Map<String, RealScope> children = new LinkedHashMap<String, RealScope>();
+  protected final Map<String, RealScope> children = new LinkedHashMap<>();
 
   protected boolean dead;
 
-  private final Set<Scoped> tearDowns = new HashSet<Scoped>();
+  private final Set<Scoped> tearDowns = new HashSet<>();
   private final ObjectGraph graph;
   private final RealScope parent;
   private final String name;
@@ -61,10 +64,11 @@ class RealScope implements MortarScope {
 
   @Override public void register(Scoped scoped) {
     if (scoped instanceof Bundler) {
-      throw new IllegalArgumentException(String.format("Scope %s cannot register %s instance %s. "
-          + "Only %ss and their children can provide bundle services", getName(),
+      throw new IllegalArgumentException(format("Scope %s cannot register %s instance %s. "
+              + "Only %ss and their children can provide bundle services", getName(),
           Bundler.class.getSimpleName(), ((Bundler) scoped).getMortarBundleKey(),
-          MortarActivityScope.class.getSimpleName()));
+          MortarActivityScope.class.getSimpleName()
+      ));
     }
 
     doRegister(scoped);
@@ -72,11 +76,15 @@ class RealScope implements MortarScope {
 
   void doRegister(Scoped scoped) {
     assertNotDead();
-    tearDowns.add(scoped);
+    if (tearDowns.add(scoped)) scoped.onEnterScope(this);
   }
 
   RealScope getParent() {
     return parent;
+  }
+
+  boolean isRoot() {
+    return parent == null;
   }
 
   @Override public RealScope findChild(String childName) {
@@ -113,6 +121,15 @@ class RealScope implements MortarScope {
     return new MortarContextWrapper(parentContext, this);
   }
 
+  @Override public void destroyChild(MortarScope child) {
+    String name = child.getName();
+    RealScope realChild = children.get(name);
+    if (realChild != child) {
+      throw new IllegalArgumentException(format("%s was created by another scope", name));
+    }
+    realChild.doDestroy();
+  }
+
   void replaceChild(String childName, RealScope scope) {
     if (scope.getParent() != this) {
       throw new IllegalArgumentException("Replacement scope must have receiver as parent");
@@ -124,20 +141,20 @@ class RealScope implements MortarScope {
     children.remove(child.getName());
   }
 
-  @Override public void destroy() {
+  void doDestroy() {
     if (dead) return;
     dead = true;
 
-    for (Scoped s : tearDowns) s.onDestroy();
+    for (Scoped s : tearDowns) s.onExitScope();
     tearDowns.clear();
     if (parent != null) parent.onChildDestroyed(this);
 
-    List<MortarScope> snapshot = new ArrayList<MortarScope>(children.values());
-    for (MortarScope child : snapshot) child.destroy();
+    List<RealScope> snapshot = new ArrayList<>(children.values());
+    for (RealScope child : snapshot) child.doDestroy();
   }
 
   @Override public String toString() {
-    return "RealMortarScope@" + Integer.toHexString(System.identityHashCode(this)) + "{" +
+    return "RealScope@" + toHexString(System.identityHashCode(this)) + "{" +
         "name='" + getName() + '\'' +
         '}';
   }