+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+Version 0.15 *(TBD)*
+------------------
+  * API break: Presenter#onDestroy and Scoped#onDestroy are now onExitScope(MortarScope).
+    Also adds onEnterScope(MortarScope) to those classes.
+
+  * API break: MortarScope#destroyChild(MortarScope) replaces MortarScope#destroy.
+
 Version 0.14 *(2014-04-18)*
 ------------------
   * Refine deferral of calls to Bundler#onLoad from MortarActivityScope#onRegister.